+++ neoforge/src/main/java/io/github/mortuusars/envelope/neoforge/datagen/server/ItemTagsDatagen.java
@@ -2,10 +2,8 @@ package io.github.mortuusars.envelope.neoforge.datagen.server;
 
 import io.github.mortuusars.envelope.Envelope;
 import net.minecraft.core.HolderLookup;
-import net.minecraft.core.registries.Registries;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.tags.ItemTagsProvider;
-import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.ItemTags;
 import net.minecraft.world.item.Items;